+++ backend/app/auth/routes/__init__.py
@@ -1,0 +1,4 @@
+from .signup import *
+from .login import *
+from .logout import *
+from .user import *
