+++ backend/app/main/routes/home.py
@@ -1,0 +1,6 @@
+from flask import jsonify
+from app.main import main
+
+@main.route('/')
+def index():
+    return jsonify({"message": "Hi from the FijaBoard API 👋"})
